requirements.txt
# Core Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
python-dotenv==1.0.0

# API Clients
requests==2.31.0
urllib3==2.0.4
certifi==2023.7.22
groq==0.4.1
stability-sdk==0.8.0
wikipedia-api==0.6.0
SPARQLWrapper==2.0.0
gnews==0.3.6

# Data Processing
numpy==1.24.4
pandas==2.0.3
pillow==10.0.0
pydantic==2.3.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Visualization
plotly==5.16.1
matplotlib==3.7.3

# Caching & Optimization
redis==4.6.0
werkzeug-cache==0.3.0

# Production
gunicorn==21.2.0
gevent==23.9.1

# Development Tools
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.1.0
python-decouple==3.8
